pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Url shortener with QrCode generator."
readme = "README.md"
requires-python = ">=3.13.11"
dependencies = [
    "aerich[toml]>=0.9.2",
    "cryptography>=46.0.7",
    "fastapi[standard]>=0.136.0",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "qrcode[pil]>=8.2",
    "tortoise-orm[asyncmy]>=0.25.4",
]

[tool.aerich]
tortoise_orm = "database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
